ml/pyproject.toml
[project]
name = "sf-traffic-forecaster-ml"
version = "0.1.0"
description = "ML forecasting pipeline for SF traffic"
requires-python = ">=3.12"
dependencies = [
    "xgboost>=3.2.0",
    "scikit-learn>=1.8.0",
    "pandas>=3.0.1",
    "numpy>=2.4.3",
    "psycopg2-binary>=2.9.11",
    "joblib>=1.5.3",
    "click>=8.3.1",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = [".", "src"]

[tool.ruff]
line-length = 100
